refactor(components): rename component files to PascalCase
