+++ src/components/sidebar/index.tsx
@@ -2,8 +2,8 @@
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import Link from "next/link";
-import Menu from "./menu";
-import SidebarToggle from "./sidebar-toggle";
+import Menu from "./SidebarMenu";
+import SidebarToggle from "./SidebarToggle";
 import { useSidebar } from "@/store/useSidebar";
 
 const Sidebar = () => {
